// admin/src/api/model/adminUserResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * FastAPI
 * Admin api endpoints
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUserResponse {
  id: string;
  created_at: string;
  updated_at: string | null;
  first_name: string;
  last_name: string;
  email: string;
  is_admin: boolean;
  is_active: boolean;
}